refactor: store useId and dialogueId instead of full objects when creating

+++ server/src/inquiry/inquiry.controller.ts
@@ -8,6 +8,7 @@ import {
   Delete,
   UseGuards,
   Request,
+  UnauthorizedException,
 } from '@nestjs/common';
 import { InquiryService } from './inquiry.service';
 import { CreateInquiryDto } from './dto/create-inquiry.dto';
@@ -24,12 +25,16 @@ export class InquiryController {
   @Post()
   @UseGuards(AccessTokenGuard)
   @ApiOperation({
-    summary: '문의사항 작성(JWT 인증 필요)',
+    summary: '문의사항 작성(Admin JWT 인증 필요)',
     description: '문의사항을 작성합니다.',
   })
   @ApiBody({ type: CreateInquiryDto })
   create(@Request() req, @Body() createInquiryDto: CreateInquiryDto) {
-    return this.inquiryService.create(createInquiryDto, req);
+    if (!req.user.isAdmin) {
+      throw new UnauthorizedException('Only Admins are allowed to create inquiries.');
+    }
+
+    return this.inquiryService.create(req.user.id, createInquiryDto);
   }
 
   @Get()
@@ -53,20 +58,30 @@ export class InquiryController {
   @Patch(':inquiryId')
   @UseGuards(AccessTokenGuard)
   @ApiOperation({
-    summary: '해당 ID 문의사항 업데이트(JWT 인증 필요)',
+    summary: '해당 ID 문의사항 업데이트(Admin JWT 인증 필요)',
     description: '해당 ID의 문의사항을 업데이트합니다.',
   })
-  update(@Param('inquiryId') inquiryId: string, @Body() updateInquiryDto: UpdateInquiryDto) {
+  update(
+    @Request() req,
+    @Param('inquiryId') inquiryId: string,
+    @Body() updateInquiryDto: UpdateInquiryDto,
+  ) {
+    if (!req.user.isAdmin) {
+      throw new UnauthorizedException('Only Admins are allowed to update inquiries.');
+    }
     return this.inquiryService.update(inquiryId, updateInquiryDto);
   }
 
   @Delete(':inquiryId')
   @UseGuards(AccessTokenGuard)
   @ApiOperation({
-    summary: '해당 ID 문의사항 삭제(JWT 인증 필요)',
+    summary: '해당 ID 문의사항 삭제(Admin JWT 인증 필요)',
     description: '해당 ID의 문의사항을 삭제합니다.',
   })
-  remove(@Param('inquiryId') inquiryId: string) {
+  remove(@Request() req, @Param('inquiryId') inquiryId: string) {
+    if (!req.user.isAdmin) {
+      throw new UnauthorizedException('Only Admins are allowed to delete inquiries.');
+    }
     return this.inquiryService.remove(inquiryId);
   }
 }
